+++ includes/Sort_QuickSort.h
@@ -15,7 +15,7 @@ public:
     vector<int> CPP() override
     {
         vector<int> arrayToSort = getToSort();
-        quickSort(arrayToSort, 0, arrayToSort.size() - 1);
+        quickSort(arrayToSort, 0, static_cast<int>(arrayToSort.size()) - 1);
         return arrayToSort;
     }
